Add files via upload
Aerofit - Business Case Study - Descriptive Statistics
& Probability
The main objective of this case study is to provide insights about the characterstics of the targeted audience for each
type of treadmill offered Aerofit, to provide a better recommendation of the treadmills to the new customers.
About the dataset
Aerofit is a leading brand in the field of fitness equipment. Aerofit provides a product range including machines such
as treadmills, exercise bikes, gym equipment, and fitness accessories to cater to the needs of all categories of people.
The company collected the data on individuals who purchased a treadmill from the AeroFit stores during the prior
three months. The dataset has the following features:
Product Purchased: KP281, KP481, or KP781
Age: In years
Gender: Male/Female
Education: In years
MaritalStatus: Single or partnered
Usage: The average number of times the customer plans to use the treadmill each week.
Income: Annual income (in $)
Fitness: Self-rated fitness on a 1-to-5 scale, where 1 is the poor shape and 5 is the excellent shape.
Miles: The average number of miles the customer expects to walk/run each week
And below is the Product Portfolio :
KP281: an entry-level treadmill that sells for $1,500
KP481: for mid-level runners that sell for $1,750
KP781: treadmill with advanced features that sell for $2,500

Business Problem:
Analyse the dataset and generate the insights on characteristics of the target audience for each type of treadmill
offered by the company, to provide a better recommendation of the treadmills to the new customers. The team then
decides investigate whether there are differences across the product with respect to customer characteristics
Objective:
1. Perform descriptive analytics to create a customer profile for each AeroFit treadmill product by developing
appropriate tables and charts.
2. For each AeroFit treadmill product, construct two-way contingency tables and compute all conditional and
marginal probabilities along with their insights/impact on the business.